refactoring: make `Logger` reusable by both primary ADS server and future MADS server
